Make job count statistics rely on doubles rather than ints

+++ src/tarch/multicore/tbb/Jobs.h
@@ -86,7 +86,7 @@ namespace tarch {
            *
            * @see spawnBackgroundJob()
            */
-          tbb::atomic<int>         maxSize;
+          tbb::atomic<double>         maxSize;
         };
         extern JobQueue _pendingJobs[NumberOfJobQueues];
